requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.1.0
jinja2>=3.1.0
pyyaml>=6.0.0
openpyxl>=3.1.0
streamlit>=1.30.0
plotly>=5.18.0
